audiofileplayer 1.1.0 - move static background audio methods into instance methods in a singleton class AudioSystem
